+++ MenuShellDemo/MenuShellDemo.csproj
@@ -46,6 +46,7 @@
     <Compile Include="Domain\Database.cs" />
     <Compile Include="Domain\User.cs" />
     <Compile Include="Domain\View\AdministratorView.cs" />
+    <Compile Include="Domain\View\LoginView.cs" />
     <Compile Include="Domain\View\ReceptionistView.cs" />
     <Compile Include="Domain\View\VeterinarianView.cs" />
     <Compile Include="Program.cs" />
